Tests for getCTAIfValid
